JAMES-1902 renames rabbitmq mailqueue guice bindings module
