+++ server/container/guice/pom.xml
@@ -159,12 +159,6 @@
                 <artifactId>james-server-guice-pop</artifactId>
                 <version>${project.version}</version>
             </dependency>
-
-            <dependency>
-                <groupId>${james.groupId}</groupId>
-                <artifactId>james-server-guice-rabbitmq</artifactId>
-                <version>${project.version}</version>
-            </dependency>
             <dependency>
                 <groupId>${james.groupId}</groupId>
                 <artifactId>james-server-guice-sieve-file</artifactId>
@@ -254,6 +248,11 @@
             <dependency>
                 <groupId>${james.groupId}</groupId>
                 <artifactId>queue-activemq-guice</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>${james.groupId}</groupId>
+                <artifactId>queue-rabbitmq-guice</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
