+++ server/container/guice/queue/rabbitmq/pom.xml
@@ -26,7 +26,7 @@
 	<relativePath>../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>james-server-guice-rabbitmq</artifactId>
+    <artifactId>queue-rabbitmq-guice</artifactId>
 
     <name>Apache James :: Server :: Guice :: RabbitMQ</name>
     <description>Guice Module for RabbitMQ</description>
